remove some implicit convert warnings
git-svn-id: http://www.openmesh.org/svnrepo/OpenMesh/trunk@1275 fdac6126-5c0c-442c-9429-916003d36597

+++ src/OpenMesh/Tools/Subdivider/Uniform/Composite/CompositeT.cc
@@ -167,7 +167,7 @@ void CompositeT<MeshType,RealType>::Tvv4()
   typename MeshType::EdgeIter         e_it;
   typename MeshType::VertexIter       v_it;
   typename MeshType::Point            zero_point(0.0, 0.0, 0.0);
-  unsigned int                    n_edges, n_faces, n_vertices, j;
+  size_t                              n_edges, n_faces, n_vertices, j;
 
   // Store number of original edges
   n_faces    = mesh_.n_faces();
@@ -225,7 +225,7 @@ void CompositeT<MeshType,RealType>::Tfv()
   typename MeshType::FaceFaceIter     ff_it;
   typename MeshType::Point            cog;
   const typename MeshType::Point      zero_point(0.0, 0.0, 0.0);
-  unsigned int                    n_edges, n_faces, n_vertices, j, valence;
+  size_t                              n_edges, n_faces, n_vertices, j, valence;
 
   // Store number of original edges
   n_faces = mesh_.n_faces();
